Adding ApiCall class adjusting ArticleContent.java and CuratedContent.java
