+++ src/test/java/com/wikia/webdriver/testcases/articlecrudtests/ArticleFeaturesCRUDTestsUser.java
@@ -1,14 +1,10 @@
 package com.wikia.webdriver.testcases.articlecrudtests;
 
-import org.joda.time.DateTime;
-import org.testng.annotations.Test;
-
 import com.wikia.webdriver.common.contentpatterns.PageContent;
 import com.wikia.webdriver.common.contentpatterns.VideoContent;
-import com.wikia.webdriver.common.core.ArticleContent;
-import com.wikia.webdriver.common.core.TestContext;
 import com.wikia.webdriver.common.core.annotations.Execute;
 import com.wikia.webdriver.common.core.annotations.User;
+import com.wikia.webdriver.common.core.api.ArticleContent;
 import com.wikia.webdriver.common.dataprovider.ArticleFeaturesCRUDDataProvider;
 import com.wikia.webdriver.common.templates.NewTestTemplate;
 import com.wikia.webdriver.pageobjectsfactory.componentobject.addphoto.AddPhotoComponentObject;
@@ -32,6 +28,9 @@ import com.wikia.webdriver.pageobjectsfactory.pageobject.article.ArticlePageObje
 import com.wikia.webdriver.pageobjectsfactory.pageobject.article.editmode.SourceEditModePageObject;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.article.editmode.VisualEditModePageObject;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.article.editmode.VisualEditModePageObject.Components;
+
+import org.joda.time.DateTime;
+import org.testng.annotations.Test;
 
 /**
  * @author Karol 'kkarolk' Kujawiak
@@ -67,7 +66,7 @@ public class ArticleFeaturesCRUDTestsUser extends NewTestTemplate {
   @Execute(asUser = User.USER)
   public void modifyGallery() {
     ArticleContent.push("<gallery position=\"right\" columns=\"1\" spacing=\"medium\">\n"
-        + "Image010.jpg\n" + "Image009.jpg\n" + "</gallery>");
+                        + "Image010.jpg\n" + "Image009.jpg\n" + "</gallery>");
 
     VisualEditModePageObject visualEditMode = new VisualEditModePageObject(driver).open();
     GalleryBuilderComponentObject galleryBuiler =
@@ -89,7 +88,8 @@ public class ArticleFeaturesCRUDTestsUser extends NewTestTemplate {
   @Execute(asUser = User.USER)
   public void deleteGallery() {
     ArticleContent.push("<gallery position=\"right\" columns=\"2\" spacing=\"medium\">\n"
-        + "Image010.jpg\n" + "Image009.jpg\n" + "Image008.jpg\n" + "Image007.jpg\n" + "</gallery>");
+                        + "Image010.jpg\n" + "Image009.jpg\n" + "Image008.jpg\n" + "Image007.jpg\n"
+                        + "</gallery>");
 
     VisualEditModePageObject visualEditMode = new VisualEditModePageObject(driver).open();
     visualEditMode.removeComponent(Components.GALLERY);
@@ -286,7 +286,8 @@ public class ArticleFeaturesCRUDTestsUser extends NewTestTemplate {
       dataProvider = "getTableProperties", groups = {"ArticleFeaturesCRUDUser_011"})
   @Execute(asUser = User.USER)
   public void ArticleFeaturesCRUDUser_011_addingTable(int border, int width, int height,
-      int cellspacing, int cellpadding, Alignment alignment) {
+                                                      int cellspacing, int cellpadding,
+                                                      Alignment alignment) {
     ArticlePageObject article = new ArticlePageObject(driver).openRandomArticle(wikiURL);
     VisualEditModePageObject visualEditMode = article.goToCurrentArticleEditPage();
     visualEditMode.clearContent();
@@ -315,13 +316,14 @@ public class ArticleFeaturesCRUDTestsUser extends NewTestTemplate {
       enabled = false)
   @Execute(asUser = User.USER)
   public void ArticleFeaturesCRUDUser_012_modifyTable(int border, int width, int height,
-      int cellspacing, int cellpadding, Alignment alignment) {
+                                                      int cellspacing, int cellpadding,
+                                                      Alignment alignment) {
     ArticlePageObject article = new ArticlePageObject(driver).openRandomArticle(wikiURL);
     SourceEditModePageObject sourceEditMode = article.openCurrectArticleSourceMode();
     sourceEditMode.clearSource();
     String table =
         sourceEditMode.buildTablePropertiesContent(border, width, height, cellspacing, cellpadding,
-            alignment);
+                                                   alignment);
     sourceEditMode.addContent(table);
     sourceEditMode.submitArticle();
     article.verifyTableBorder(border);
@@ -344,13 +346,14 @@ public class ArticleFeaturesCRUDTestsUser extends NewTestTemplate {
       dataProvider = "getTableProperties", groups = {"ArticleFeaturesCRUDUser_013"})
   @Execute(asUser = User.USER)
   public void ArticleFeaturesCRUDUser_013_deleteTable(int border, int width, int height,
-      int cellspacing, int cellpadding, Alignment alignment) {
+                                                      int cellspacing, int cellpadding,
+                                                      Alignment alignment) {
     ArticlePageObject article = new ArticlePageObject(driver).openRandomArticle(wikiURL);
     SourceEditModePageObject sourceEditMode = article.openCurrectArticleSourceMode();
     sourceEditMode.clearSource();
     String table =
         sourceEditMode.buildTablePropertiesContent(border, width, height, cellspacing, cellpadding,
-            alignment);
+                                                   alignment);
     sourceEditMode.addContentInSourceMode(table);
     sourceEditMode.submitArticle();
     VisualEditModePageObject visualEditMode = article.goToCurrentArticleEditPage();
